+++ src/pages/admin/Outgoing.jsx
@@ -18,6 +18,8 @@ const Outgoing = () => {
   const [selectedYear, setSelectedYear] = useState("All");
   const [selectedMonth, setSelectedMonth] = useState("All");
   const [currentPage, setCurrentPage] = useState(1);
+    const [setOpen, setIsOpen] = useState(false);
+    const [selectedDocId, setSelectedDocId] = useState(null);
   const itemsPerPage = 15;
 
   const fetchDocument = async () => {
@@ -85,12 +87,17 @@ const Outgoing = () => {
   ];
 
   const deleteDocument = async (id) => {
+    if (!id) {
+      console.error("Error: Document ID is null or undefined");
+      return;
+    }
     try {
       await axios.delete(
         `http://localhost:5000/api/document/delete-document/${id}`
       );
-      setOutgoing(outgoing.filter((doc) => doc._id !== id));
-      setFilteredDocs(filteredDocs.filter((doc) => doc._id !== id));
+      setOutgoing((prev) => prev.filter((doc) => doc._id !== id));
+      setFilteredDocs((prev) => prev.filter((doc) => doc._id !== id));
+      setIsOpen(false);
     } catch (error) {
       console.error("Error deleting document:", error);
     }
@@ -456,7 +463,10 @@ const Outgoing = () => {
                             Edit
                           </button>
                           <button
-                            onClick={() => deleteDocument(doc._id)}
+                             onClick={() => {
+        setSelectedDocId(doc._id);
+        setIsOpen(true);
+      }}
                             className="bg-red-500 text-white px-2 py-1 rounded"
                           >
                             Delete
@@ -468,6 +478,30 @@ const Outgoing = () => {
                 ))}
               </tbody>
             </table>
+            <div className="bg-gray-500 ">
+  {setOpen && (
+    <div className="fixed inset-0 flex items-center justify-center bg-gray-950/70 backdrop-blur-50 z-50">
+        <div className="bg-white p-6 rounded-lg shadow-lg relative z-10 w-90">
+          <p className="font-medium">Are you sure you want to delete?</p>
+          <p>This will permanently delete the item and remove it from our servers.</p>
+          <div className="w-full flex justify-end items-end space-x-2">
+            <button
+              onClick={() => deleteDocument(selectedDocId)}
+              className="bg-red-500 text-white px-4 py-2 w-20 rounded mt-4"
+            >
+              Yes
+            </button>
+            <button
+              onClick={() => setIsOpen(false)}
+              className="bg-white text-black border border-gray-400 px-4 py-2 rounded mt-4"
+            >
+              Cancel
+            </button>
+          </div>
+        </div>
+      </div>
+  )}
+</div>
           </>
         ) : (
           <p className="text-center text-gray-500 mt-4">
